requirements.txt
asgiref==3.5.2
Django==4.1.4
djangorestframework==3.14.0
mysqlclient==2.1.1
pytz==2022.7
sqlparse==0.4.3
